sensor-code/requirements.txt
Adafruit-Blinka==8.32.0
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-requests==2.0.5
adafruit-circuitpython-typing==1.10.2
adafruit-circuitpython-vl53l0x==3.6.11
Adafruit-PlatformDetect==3.60.0
Adafruit-PureIO==1.1.11
board==1.0
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
numpy==1.26.4
pyftdi==0.55.0
pyserial==3.5
pyusb==1.2.1
requests==2.31.0
smbus2==0.4.3
typing_extensions==4.9.0
urllib3==2.2.0
RPI.GPIO==0.7.1
